Refactor profiles/serializers.py into a module folder, in preparation for an additional OrganizationProfile View
