+++ pulseapi/profiles/models/__init__.py
@@ -1,5 +1,5 @@
-from .profiletype import ProfileType
-from .programtype import ProgramType
-from .programyear import ProgramYear
-from .userprofile import UserProfile, entry_thumbnail_path
-from .userbookmarks import UserBookmarks
+from pulseapi.profiles.models.profiletype import ProfileType
+from pulseapi.profiles.models.programtype import ProgramType
+from pulseapi.profiles.models.programyear import ProgramYear
+from pulseapi.profiles.models.userprofile import UserProfile, entry_thumbnail_path
+from pulseapi.profiles.models.userbookmarks import UserBookmarks
