+++ pulseapi/profiles/serializers/__init__.py
@@ -1,0 +1,8 @@
+# from pulseapi.profiles.serializers.userprofile import *
+
+from profiles.serializers.userprofile.serializer import UserProfileSerializer
+from profiles.serializers.userprofile.basic import UserProfileBasicSerializer
+from profiles.serializers.userprofile.public import UserProfilePublicSerializer
+from profiles.serializers.userprofile.entries import UserProfileEntriesSerializer
+from profiles.serializers.userprofile.public_entries import UserProfilePublicWithEntriesSerializer
+from profiles.serializers.userprofile.userbookmarks import UserBookmarksSerializer
