+++ pulseapi/profiles/serializers/userprofile/basic.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+
+class UserProfileBasicSerializer(serializers.BaseSerializer):
+    """
+    A read-only serializer that serializes a user profile by only including indentity
+    information like `id` and `name`.
+    """
+    def to_representation(self, obj):
+        return {
+            'id': obj.id,
+            'name': obj.name
+        }
